refactor(Application, GameControllerFactory): GameControllerFactory 를 controller 에서 config 로 이동해 패키지마다의 책임을 명확히 해라
